feat(rfc-0005): Fase 2 — migra commands para src/hronir/commands.ts
2235 linhas convertidas para TypeScript com tipos explícitos nas
assinaturas exportadas e nas estruturas de dados internas.
